refactor: group files into directories
