+++ src/App.tsx
@@ -1,11 +1,11 @@
-import { MerkleTreeView } from "./MerkleTreeView.tsx";
-import { MerkleTreeProvider } from "./MerkleTreeProvider.tsx";
-import { MerkleTreeSizeSelector } from "./MerkleTreeSizeSelector.tsx";
-import { MerkleProofProvider } from "./MerkleProofsProvider.tsx";
-import { MerkleProofsView } from "./MerkleProofsView.tsx";
-import { ResetProvider } from "./useReset.tsx";
-import { TscMerkleProofsView } from "./TscMerkleProofsView.tsx";
-import { BsvUnifiedMerklePathView } from "./BsvUnifiedMerklePathView.tsx";
+import { MerkleTreeView } from "./components/MerkleTreeView/MerkleTreeView.tsx";
+import { MerkleTreeProvider } from "./providers/MerkleTreeProvider/MerkleTreeProvider.tsx";
+import { MerkleTreeSizeSelector } from "./components/MerkleTreeSizeSelector/MerkleTreeSizeSelector.tsx";
+import { MerkleProofProvider } from "./providers/MerkleProofsProvider/MerkleProofsProvider.tsx";
+import { MerkleProofsView } from "./components/MerkleProofsView/MerkleProofsView.tsx";
+import { ResetProvider } from "./providers/ResetStateProvider/ResetStateProvider.tsx";
+import { TscMerkleProofsView } from "./components/TscMerkleProofsView/TscMerkleProofsView.tsx";
+import { BsvUnifiedMerklePathView } from "./components/BsvUnifiedMerklePathView/BsvUnifiedMerklePathView.tsx";
 import { Grid } from "@mui/material";
 
 function App() {
