+++ src/types/merkle-tree-data.d.ts
